% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dgammadf}
\alias{dgammadf}
\title{Density of matrix gamma distribution with vector degrees of freedom}
\usage{
dgammadf(X, a = rep(1, 2), A = diag(length(a)))
}
\arguments{
\item{X}{matrix-point, argument for density function (r x r)}

\item{a}{vector of degrees of freedom for underlying gamma Bellman distribution (r x 1)}

\item{A}{scale matrix for underlying gamma Bellman distribution (r x r)}
}
\value{
scalar, value of density function
}
\description{
Density of matrix gamma distribution with vector degrees of freedom
}
\details{
Density of matrix gamma distribution with vector degrees of freedom. The distribution
is described in Shvedov
generalisation of Algorithm by Alexey Balaev
}
\examples{
d <- dgammadf(X = matrix(1, nrow=2, ncol=2), a=c(1,2))
d
}

