% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dmatnorm}
\alias{dmatnorm}
\title{Matrix Normal density function}
\usage{
dmatnorm(X, Mu = matrix(0), U = diag(nrow(Mu)), V = diag(ncol(Mu)))
}
\arguments{
\item{X}{matrix-point, argument for density function}

\item{Mu}{matrix of expected values (r x s)}

\item{U}{among-row scale covariance matrix (r x r)}

\item{V}{among-column scale covariance matrix (s x s)}
}
\value{
scalar, density at the point X
}
\description{
Matrix Normal density function
}
\details{
Matrix Normal density function
}
\examples{
d <- dmatnorm(X = matrix(1, nrow=3, ncol=2), Mu = matrix(0, nrow=3, ncol=2))
d
}

