% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dtdf}
\alias{dtdf}
\title{Density of matrix t distribution with vector degrees of freedom}
\usage{
dtdf(X, M = matrix(0), B = diag(ncol(M)), a = rep(1, nrow(M)),
  A = diag(nrow(M)))
}
\arguments{
\item{X}{matrix-point, argument for density function}

\item{M}{matrix of expected values (r x s)}

\item{B}{among-column scale covariance matrix for underlying standard normal (s x s)}

\item{a}{vector of degrees of freedom for underlying gamma Bellman distribution (r x 1)}

\item{A}{scale matrix for underlying gamma Bellman distribution (r x r)}
}
\value{
array of generated matrix t (n x r x s)
}
\description{
Density of matrix t distribution with vector degrees of freedom
}
\details{
Density of matrix t distribution with vector degrees of freedom. The distribution
is described in Shvedov
generalisation of Algorithm by Alexey Balaev
}
\examples{
d <- dtdf(X = matrix(1, nrow=3, ncol=2), M = matrix(0, nrow=3, ncol=2))
d
}

