% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rgammadf}
\alias{rgammadf}
\title{Generate gamma distribution with vector degrees of freedom}
\usage{
rgammadf(n = 1, a = rep(1, 2), A = diag(length(a)))
}
\arguments{
\item{n}{number of draws from matrix gamma distribution with vector degrees of freedom}

\item{a}{vector of degrees of freedom (r x 1)}

\item{A}{scale matrix (r x r)}
}
\value{
array of generated matrix gamma (n x r x r)
}
\description{
Generate gamma distribution with vector degrees of freedom
}
\details{
Generate gamma distribution with vector degrees of freedom. The distribution
is described in Shvedov
Algorithm by Alexey Balaev
}
\examples{
X <- rgammadf(n = 5, Mu = matrix(0, nrow=3, ncol=2))
X
}

