% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dtdf}
\alias{dtdf}
\title{Density of matrix t distribution with vector degrees of freedom}
\usage{
dtdf(X, M = matrix(0, nrow = length(a), ncol = ncol(B)), B = diag(1),
  a = rep(1, 2), A = diag(length(a)))
}
\arguments{
\item{X}{matrix-point, argument for density function}

\item{M}{matrix of expected values (r x s)}

\item{B}{among-column scale covariance matrix for underlying standard normal (s x s)}

\item{a}{vector of degrees of freedom for underlying gamma Bellman distribution (r x 1)}

\item{A}{scale matrix for underlying gamma Bellman distribution (r x r)}
}
\value{
scalar, value of density function
}
\description{
Density of matrix t distribution with vector degrees of freedom
}
\details{
Density of matrix t distribution with vector degrees of freedom. The distribution
is described in Shvedov
generalisation of Algorithm by Alexey Balaev
}
\examples{
d <- dtdf(X = matrix(1, nrow=3, ncol=1), a = c(1, 2, 3))
d
}

