% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mc_sim}
\alias{mc_sim}
\title{Generate Discrete Markov Chain}
\usage{
mc_sim(trans, N = 1, start = 1)
}
\arguments{
\item{trans}{square transition matrix}

\item{N}{number of states to generate}

\item{start}{either a starting state (default is state 1) or a vector of probabilities for starting state}
}
\value{
integer vector of simulated states
}
\description{
Generate Discrete Markov Chain
}
\details{
Generate Discrete Markov Chain given transition matrix
}
\examples{
P <- 0.6*diag(3) + matrix(0.2, nrow=3, ncol=3)
mc <- mc_sim(P, N=100)
mc
}

